Create break_cycles
You are given an undirected multi-graph. Remove a minimal number of edges to make the graph acyclic (to break all cycles). As a result, print the number of edges removed and the removed edges. If several edges can be removed to break a certain cycle, remove the smallest of them in alphabetical order (smallest start vertex in alphabetical order and smallest end vertex in alphabetical order).